Integrate S2H to shaders
Add basic hello world text to my RT render + box rasterizer. Making a new shader for the box with S2H since it can only print inside the box and don't want this to be part of my current shader.
TODO: Add background in box render as part of shader and not imgui so it can display S2H